requirements.txt
openai>=1.3.5
spacy>=3.7.2
python-docx>=1.1.0
pdf2image>=1.16.3
pytesseract>=0.3.10
PyMuPDF>=1.23.9
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
aiofiles>=23.2.1

# Optional: For webhook support
requests>=2.31.0

# Optional: For IMAP-based email polling
imapclient>=2.3.1
pyzmail36>=1.0.4
